feat(profile): Add commands for managing user profile titles
Rename the `/profil` command to `/profil user` to enable adding more
commands to the `/profil` group.

Add moderator commands under the `/profil-tytuły` group:
- Managing profile titles
- Viewing all titles
- Assigning titles to users

Add user commands under the `/profil tytuł` group:
- Viewing owned titles
- Selecting the active title

Display the selected for the current user in `/profil user` and default
to an empty title if none is set.