fix(ultimatum): Always update `endedAt` when an ultimatum ends
This makes the ultimatum be marked as "ended" in our DB, even when the
user isn't in the guild at the time it ends.

This is a little confusing in terms of logging, becase we'll remove the
ultimatum without logging anything, but it's better to have an ultimatum
marked as done in the DB than to keep indefinitely.

Fixes #133